+++ src/main/java/nl/lexemmens/podman/enumeration/SkopeoCommand.java
@@ -1,0 +1,17 @@
+package nl.lexemmens.podman.enumeration;
+
+public enum SkopeoCommand {
+
+    SKOPEO("skopeo"),
+    COPY("copy");
+
+    private String command;
+
+    SkopeoCommand(String command) {
+        this.command = command;
+    }
+
+    public String getCommand() {
+        return command;
+    }
+}